requirements.txt
matplotlib==3.9.2
moviepy==2.0.0.dev2
numpy==2.1.2
opencv_python==4.10.0.84
Pillow==10.4.0
proglog==0.1.10
PyQt5==5.15.11
PyQt5_sip==12.15.0
seaborn==0.13.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
